Just tried to test your time manipulation process
Technically, i didn't find a way to create an actual Date-Time object in C++ so what i'm doing is that i'm initializing it with today's system Date-Time (which is the only use i found online) and then setting its attributes like you did with the user input. This is misleading because the tm structure has other attributes we're not setting properly- for example, we're not changing the day of the week, and so it will always be the system's initialised day. To not make this problem obvious, you can convert tm to string with strftime() instead of asctime() and so format it like you want (displaying only what you need). Ref. https://www.geeksforgeeks.org/strftime-function-in-c/
One last notice, i prefer that you take the user input in one string and check if it's in the correct format (ddmmyyyy hh:mm) in a while loop then manipulate that string (instead of taking in every bit alone).
NB:  int tm_year is the year since 1900.